+++ src/integration-test/groovy/org/pih/warehouse/api/spec/core/LocalizationApiSpec.groovy
@@ -1,27 +1,16 @@
 package org.pih.warehouse.api.spec.core
 
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.test.context.TestPropertySource
 
 import org.pih.warehouse.api.client.core.LocalizationApiWrapper
 import org.pih.warehouse.api.spec.base.ApiSpec
-import org.pih.warehouse.core.Localization
 import org.pih.warehouse.core.localization.LocalizedMessageDto
 import org.pih.warehouse.core.localization.LocalizedMessagesDto
 
-@TestPropertySource(properties = ["openboxes.locale.custom.enabled=true"])
 class LocalizationApiSpec extends ApiSpec {
 
     @Autowired
     LocalizationApiWrapper localizationApiWrapper
-
-    void setupData() {
-        Localization.findOrBuild(
-                code: 'default.lot.label',
-                locale: 'fr',
-                text: 'overridden',
-        )
-    }
 
     void 'listing the messages returns without error for all supported locales'() {
         when: 'we fetch the default locale messages'
@@ -59,15 +48,5 @@ class LocalizationApiSpec extends ApiSpec {
         assert message.code == 'default.lot.label'
         assert message.message == 'Lote'
         assert message.currentLocale == new Locale('es')
-    }
-
-    void 'getting a message that is overridden successfully returns the overridden message'() {
-        when: 'we fetch a message that is overridden in the given locale'
-        LocalizedMessageDto message = localizationApiWrapper.getOK('default.lot.label', 'fr')
-
-        then: 'the message is localized to the requested locale (Spanish)'
-        assert message.code == 'default.lot.label'
-        assert message.message == 'overridden'
-        assert message.currentLocale == Locale.FRENCH
     }
 }
